x86_64: fix get_kaslr_offset_x86_64() to return kaslr_offset correctly
Currently, the get_kaslr_offset_x86_64() function has the following
condition to return info->kaslr_offset, but that is wrong and can
return 0 for kernel symbols if info->kaslr_offset is small.

    if (vaddr >= __START_KERNEL_map &&
                    vaddr < __START_KERNEL_map + info->kaslr_offset)

Consequently, symbols in erase config could be resolved wrongly, and
makedumpfile fails to vtop with the following message or erases wrong
address data.

    __vtop4_x86_64: Can't get a valid pmd_pte.

To fix this, use NUMBER(KERNEL_IMAGE_SIZE) in vmcoreinfo if any, or
a hard-coded value, which has not been changed from initial KASLR
implementation.

Signed-off-by: Kazuhito Hagio <k-hagio@ab.jp.nec.com>