[PATCH] Introduce --check-params option
Currently it's difficult to check whether a makedumpfile command-line
is valid or not without an actual panic.  This is inefficient and if
a wrong configuration is not tested, you will miss the vmcore when an
actual panic occurs.

In order for kdump facilities like kexec-tools to be able to check
the specified command-line parameters in advance, introduce the
--check-params option that only checks them and exits immediately.

Signed-off-by: Kazuhito Hagio <k-hagio-ab@nec.com>